Add files via upload
Pluses and Minuses:
The n consecutive integers from 1 to n are written in a row. Design an 
algorithm that puts signs “+” and “−” in front of them so that the expression 
obtained is equal to 0 or, if the task is impossible to do, returns the 
message “no solution.” Your algorithm should be much more efficient than 
an examination of all possible ways to place the signs.